# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/github/webhooks": {
      "post": {
        "operationId": "receiveWebhook",
        "parameters": [
          {
            "name": "x-github-delivery",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-github-event",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-hub-signature-256",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "201": {
            "description": ""
          }
        },
        "tags": [
          "github"
        ]
      }
    },
...
  },
  "info": {
    "title": "ReviewTime",
    "description": "The ReviewTime main API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
...
}